+++ jest.config.cjs
@@ -10,8 +10,8 @@ module.exports = {
     [
       'jest-html-reporter',
       {
-        pageTitle: 'QuizPulse — Sprint 5 Test Report',
-        outputPath: 'tests/reports/sprint5-report.html',
+        pageTitle: 'QuizPulse — Sprint 6 Test Report',
+        outputPath: 'tests/reports/sprint6-report.html',
         includeFailureMsg: true,
         includeConsoleLog: true,
       },